refactoring, some code cleanup around comments and console logs mostly

+++ controllers/user.js
@@ -1,9 +1,9 @@
 const express = require('express')
 const router = express.Router()
 const User = require('../models/user')
-const Favorite = require('../models/favorites')
+const Favorite = require('../models/favorite')
 const bcryptjs = require('bcryptjs')
-const Jaunt = require('../models/jaunts')
+const Jaunt = require('../models/jaunt')
 
 //user index route
 // should we get rid of this, or only make it accessible to admins maybe?
@@ -43,22 +43,20 @@ router.post('/login', async (req, res, next) => {
 		const foundUsers = await User.find({
 			username: req.body.username
 		})
-
+		// testing if the username is in the database
 		if(foundUsers.length === 0) {
-			console.log('Username Does Not Exist');
 			req.session.message = 'Invalid username or password'
 			res.redirect('/users/login')
 		} else {
 			const pw = req.body.password
-			console.log(foundUsers[0])
+			// hashing the password
 			if(bcryptjs.compareSync(pw, foundUsers[0].password)) {
 				req.session.loggedIn = true
 				req.session.username = foundUsers[0].username
 				req.session.userId = foundUsers[0]._id
-				console.log(req.session,' session info')
 				res.redirect('/users/'+req.session.userId)
 			} else {
-				console.log('bad password');
+				// if the password doesn't match hashed password in database
 				req.session.message = 'Invalid username or password'
 				res.redirect('/users/login')
 			}
@@ -79,20 +77,17 @@ router.get('/register', (req, res, next) => {
 	}
 })
 
-//Registration Route - post
+//Registration Route - put
 router.post('/', async (req, res, next) => {
 	try {
 		const userSearch = await User.findOne({username: req.body.username})
-		console.log('\nhere is what we found when trying to see if user name exists POST /users')
-		console.log(userSearch)
+		// checking if username is taken already, redirecting if so.
 		if (userSearch !== null) {
-			console.log('username is taken')
 			req.session.message = 'Username is taken!'
 			res.redirect('/')
 		} else {
 			const pw = req.body.password
 			const hashedPassword = bcryptjs.hashSync(pw, bcryptjs.genSaltSync(10))
-			console.log(hashedPassword)
 			const createdUser = await User.create({username: req.body.username, password: hashedPassword})
 			console.log('\nhere is the created user in POST /users')
 			console.log(createdUser)
@@ -110,14 +105,13 @@ router.post('/', async (req, res, next) => {
 
 // user logout route
 router.get('/logout', (req, res) => {
-
-	req.session.destroy((err) => {
-		if (err) {
-			console.log(err)
-		} else {
-			res.redirect('/')
-		}
-	})
+	// destroy session at logout
+	try {
+		req.session.destroy()		
+		res.redirect('/')
+	} catch(err) {
+		next(err)
+	}
 })
 
 
@@ -125,9 +119,7 @@ router.get('/logout', (req, res) => {
 router.get('/:id', async (req, res, next) => {
 	if (req.session.loggedIn) {
 		try {
-/*			const foundUser = await User.findById(req.params.id)*/
 			const foundFaves = await Favorite.find({user: req.session.userId})
-			console.log(foundFaves)
 			const foundJaunts = await Jaunt.find({user: req.session.userId})
 			console.log(foundJaunts)
 			res.render('users/show.ejs', {
@@ -145,18 +137,24 @@ router.get('/:id', async (req, res, next) => {
 
 //user delete route
 router.delete('/:id', async (req, res, next) => {
-	try {
-		// gotta remove affiliated faves
+	if (req.session.loggedIn){
+		try {
+			// gotta remove affiliated faves
+	
+			const deletedFaves = await Favorite.deleteMany({user: req.params.id})
+			console.log('\n affiliated faves with this user to be deleted ', deletedFaves)
+			const deleteUser = await User.deleteOne({
+				_id: req.params.id})
+			res.redirect('/users/login')
+		}
+		catch (err) {
+			next(err)
+		}
+	} else {
+			req.session.message = 'You must be logged in to do that'
+			res.redirect('/users/login')
+	}
 
-		const deletedFaves = await Favorite.deleteMany({user: req.params.id})
-		console.log('\n affiliated faves with this user to be deleted ', deletedFaves)
-		const deleteUser = await User.deleteOne({
-			_id: req.params.id})
-		res.redirect('/users/login')
-	}
-	catch (err) {
-		next(err)
-	}
 })
